BillSplit/backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
python-dotenv==1.0.0
PyJWT==2.8.0
firebase-admin==6.2.0
Werkzeug==2.3.7  # Pinning Werkzeug to avoid conflicts with Flask 2.3.2
Pydantic==2.7.1